Modularizando codigo y agregando la funcion leveling
