+++ Juego_Practica/src/datosTXT/PlayersCharacters/Tamtous.txt
@@ -1,3 +1,4 @@
 Tamtous,1,Adri,Warrior,1,0
 Tamtous,2,Franco,Warrior,1,0
 Tamtous,3,puta,Rogue,1,0
+Tamtous,4,Laura,Warrior,1,0
